feat(ckg-api): stricter CKG result match
